+++ src/ArduinoJson/DynamicJsonDocument.hpp
@@ -37,7 +37,7 @@ class DynamicJsonDocument {
 
   void clear() {
     _memoryPool.clear();
-    _rootData.setNull();
+    _rootData.type = Internals::JSON_NULL;
   }
 
   size_t memoryUsage() const {