+++ src/ArduinoJson/MsgPack/MsgPackDeserializer.hpp
@@ -277,13 +277,12 @@ class MsgPackDeserializer {
     if (_nestingLimit == 0) return DeserializationError::TooDeep;
     --_nestingLimit;
     for (; n; --n) {
-      JsonVariantData keyData;
-      JsonVariant key(_memoryPool, &keyData);
+      JsonVariantLocal key(_memoryPool);
       DeserializationError err = parse(key);
       if (err) return err;
-      if (!keyData.isString()) return DeserializationError::NotSupported;
+      if (!key.is<char *>()) return DeserializationError::NotSupported;
 
-      JsonVariant value = object.set(keyData.asString());
+      JsonVariant value = object.set(key.as<char *>());
       if (value.isInvalid()) return DeserializationError::NoMemory;
 
       err = parse(value);